Test refactor: created Singleton vs Instance mixins.
